+++ lib/fiber/expression.hpp
@@ -40,8 +40,8 @@ public:
 
     virtual ~Expression() {}
 
-    virtual size_t domainDimension() const = 0;
-    virtual size_t codomainDimension() const = 0;
+    virtual int domainDimension() const = 0;
+    virtual int codomainDimension() const = 0;
 
     virtual void addDependencies(size_t& basisDeps, size_t& geomDeps) const = 0;
