+++ lib/fiber/scalar_function_value.hpp
@@ -38,11 +38,11 @@ class ScalarFunctionValue : public Expression<CoordinateType>
 public:
     typedef typename Expression<CoordinateType>::ComplexType ComplexType;
 
-    virtual size_t domainDimension() const {
+    virtual int domainDimension() const {
         return 1;
     }
 
-    virtual size_t codomainDimension() const {
+    virtual int codomainDimension() const {
         return 1;
     }
